splits out action creators to individual files, and adds tests for config actionCreators
