+++ __tests__/src/actions/canvas.test.js
@@ -1,0 +1,25 @@
+import * as actions from '../../../src/actions/index';
+import ActionTypes from '../../../src/action-types';
+
+describe('canvas actions', () => {
+  describe('nextCanvas', () => {
+    it('moves to the next canvas', () => {
+      const id = 'abc123';
+      const expectedAction = {
+        type: ActionTypes.NEXT_CANVAS,
+        windowId: id,
+      };
+      expect(actions.nextCanvas(id)).toEqual(expectedAction);
+    });
+  });
+  describe('previousCanvas', () => {
+    it('moves to the previous canvas', () => {
+      const id = 'abc123';
+      const expectedAction = {
+        type: ActionTypes.PREVIOUS_CANVAS,
+        windowId: id,
+      };
+      expect(actions.previousCanvas(id)).toEqual(expectedAction);
+    });
+  });
+});
