+++ src/actions/canvas.js
@@ -1,0 +1,21 @@
+import ActionTypes from '../action-types';
+
+/**
+ * nextCanvas - action creator
+ *
+ * @param  {String} windowId
+ * @memberof ActionCreators
+ */
+export function nextCanvas(windowId) {
+  return { type: ActionTypes.NEXT_CANVAS, windowId };
+}
+
+/**
+ * previousCanvas - action creator
+ *
+ * @param  {String} windowId
+ * @memberof ActionCreators
+ */
+export function previousCanvas(windowId) {
+  return { type: ActionTypes.PREVIOUS_CANVAS, windowId };
+}
